Fixup kotlin warnings from Android Studio after recent upgrade
Mostly removing casts that were unneccessary and changing many var to val.
Sneak in support for audio beacon distance callout.

+++ app/src/main/java/org/scottishtecharmy/soundscape/geoengine/filters/LocationUpdateFilter.kt
@@ -56,7 +56,7 @@ open class LocationUpdateFilter(minTimeMilliseconds: Long, private val minDistan
             val timeInterval = minTimeMs * inVehicleTimeIntervalMultiplier
             var distanceInterval = minDistance
             if(userGeometry.speed > 0) {
-                distanceInterval = userGeometry.speed.toDouble() * minTimeMs
+                distanceInterval = userGeometry.speed * minTimeMs
             }
 
             return shouldUpdate(userGeometry, timeInterval, distanceInterval)
